# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organization/paths/{route_id}": {
            "get": {
                "tags": [
                    "Paths"
                ],
                "summary": "get a route's paths",
                "description": "get a route's paths",
                "operationId": "get a route's paths",
                "parameters": [
                    {
                        "name": "route_id",
                        "in": "path",
                        "description": "route_id is the route id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/organization/paths": {
            "get": {
                "tags": [
                    "Paths"
                ],
                "summary": "get all paths",
                "description": "get all paths",
                "operationId": "get all paths",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Paths"
                ],
                "summary": "add a path to a route",
                "description": "add a path to a route",
                "operationId": "add a path to a route",
                "parameters": [
                    {
                        "name": "route_id",
                        "in": "path",
                        "description": "route_id is the route id",
                        "required": true
                    },
                    {
                        "name": "path",
                        "in": "path",
                        "description": "path is a json object contains path_name and stops and each stop has name, longitude and latitude",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/organization/paths/id/{id}": {
            "get": {
                "tags": [
                    "Paths"
                ],
                "summary": "get path by id",
                "description": "get path by id",
                "operationId": "get path by id",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/organization/paths/{path_name}": {
            "put": {
                "tags": [
                    "Paths"
                ],
                "summary": "update a path",
                "description": "update a path",
                "operationId": "update a path",
                "parameters": [
                    {
                        "name": "stops",
                        "in": "path",
                        "description": "array of stops of the path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/organization/paths/{path_id}": {
            "delete": {
                "tags": [
                    "Paths"
                ],
                "summary": "delete a path",
                "description": "delete a path",
                "operationId": "delete a path",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
        "/api/organization/routes/{id}": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "get a route",
                "description": "get a route",
                "operationId": "get a route",
                "parameters": [
                    {
                        "name": "source",
                        "in": "path",
                        "description": "source of the route",
                        "required": true
                    },
                    {
                        "name": "destination",
                        "in": "path",
                        "description": "destination of the route",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id is the route id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Routes"
                ],
                "summary": "update the route",
                "description": "update the route",
                "operationId": "update the route",
                "parameters": [
                    {
                        "name": "source",
                        "in": "path",
                        "description": "source of the route",
                        "required": true
                    },
                    {
                        "name": "destination",
                        "in": "path",
                        "description": "destination of the route",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id is the route id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Routes"
                ],
                "summary": "delete the route",
                "description": "delete the route",
                "operationId": "delete the route",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id is the route id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}